chore: dummyview 삭제 후 GageButton 뷰 에셋 추가
